New refactoring code 10.04.2023 - part 2

+++ src/main/resources/liquibase/scripts/ads.sql
@@ -5,7 +5,7 @@
 CREATE TABLE IF NOT EXISTS ads
 (
     id          BIGSERIAL PRIMARY KEY,
-    author_id   integer REFERENCES users (id),
+    author_id   BIGINT REFERENCES users (id),
     price       int  NOT NULL,
     title       varchar,
     description text NOT NULL
