+++ src/main/resources/liquibase/scripts/images.sql
@@ -6,17 +6,17 @@ CREATE TABLE IF NOT EXISTS image
 (
     id         BIGSERIAL PRIMARY KEY,
     image      bytea   NOT NULL,
-    file_size  integer NOT NULL,
+    file_size  BIGINT NOT NULL,
     media_type text
 );
 
 -- changeset unit:2
 ALTER TABLE ads
-    ADD COLUMN image_id integer REFERENCES image (id);
+    ADD COLUMN image_id BIGINT REFERENCES image (id);
 
 --changeset unit:1
 ALTER TABLE image
-ADD COLUMN ads_id integer REFERENCES ads (id);
+ADD COLUMN ads_id BIGINT REFERENCES ads (id);
 
 
 --changeset unit:4
